basic button input now working - viewport raycast function implemented - raycast triangle implmeneted - added necessary math functions - fixed some issues with previous commit,   mainly ids were used as indices (added a TODO item for that)

+++ rogue/source/engine/viewport.h
@@ -31,7 +31,7 @@ namespace elem
 struct TriangleRaycastResult
 {
 	bool hit{};
-	int hit_triangle_idx{};
+	size_t hit_triangle_idx{};
 	double z{}; // z in NDC (-1, 1)
 	float w0{}, w1{}, w2{}; // barycentric coordinates
 };
